Standardise notifications file
